int/compiler/builder/propagated/bundle: moves package
