+++ internal/compiler/builder/propagated/bundle/bundle.go
